requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
pillow
numpy